Move login response classes to login package
